#625 refactor(common): Move the MQTT config
