+++ common/go/src/config/mqtt.go
@@ -1,0 +1,9 @@
+package config
+
+type MqttConfig struct {
+	Host         string
+	Port         int
+	User         string
+	PasswordFile string
+	TopicBase    string
+}
